requirements.txt
# Requirements.txt - Gelato Mágico Project

# Core ML Libraries
scikit-learn==1.0.2
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
seaborn==0.11.2
joblib==1.0.1

# MLflow and Azure ML
mlflow==2.8.1
azureml-core==1.54.0
azureml-mlflow==1.54.0
azureml-defaults==1.54.0
azureml-train-core==1.54.0

# Model deployment
inference-schema==1.5.1
flask==2.0.1

# Utilities
python-dateutil==2.8.2
pytz==2021.3
requests==2.25.1
urllib3==1.26.6

# Development and testing
pytest==6.2.4
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2

# Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.2.0
notebook==6.4.3

# Data visualization extras
plotly==5.3.1
bokeh==2.3.3

# Optional: For advanced feature engineering
scipy==1.7.1
statsmodels==0.12.2